Created "Users" tab for viewing users by administrators. The system of site navigation and verification of authorized users has been redesigned.

+++ ForumBackend/src/main/resources/database/initUsers.sql
@@ -1,9 +1,10 @@
-INSERT INTO users (city, email, avatar, login, name, phone)
-VALUES ('Samara','user1@mail.ru','file.img', 'user1','Ivan','+79023458765'),
-       ('Moscow','user2@mail.ru','file.img', 'user2','Petr','+78723458765'),
-       ('Samara','user3@mail.ru','file.img', 'user3','Diman','+79022758765');
+INSERT INTO users (username, email, password, )
+VALUES ('Vlad','user1@mail.ru','123456'),
+       ('Moscow','user2@mail.ru',''),
+       ('Samara','user3@mail.ru','');
 
 select * from users;
 select * from user_roles;
 
-DELETE from users;
+DELETE from user_roles where user_id =(select id from users where username='Ivan');
+DELETE from users where username='Ivan' ;
